+++ myreports/datasources/contacts.py
@@ -2,8 +2,8 @@
 from operator import __or__
 
 from myreports.datasources.util import (
-    DataSource, DataSourceFilter, dispatch_help_by_field_name,
-    filter_date_range)
+    dispatch_help_by_field_name, filter_date_range)
+from myreports.datasources.base import DataSource, DataSourceFilter
 
 from mypartners.models import Contact, Location, Tag, Partner, Status
