+++ myreports/datasources/partners.py
@@ -4,8 +4,8 @@ from operator import __or__
 from mypartners.models import Contact, Partner, Status, Location, Tag
 
 from myreports.datasources.util import (
-    DataSource, DataSourceFilter, dispatch_help_by_field_name,
-    filter_date_range)
+    dispatch_help_by_field_name, filter_date_range)
+from myreports.datasources.base import DataSource, DataSourceFilter
 
 
 from universal.helpers import dict_identity
